split up types and step functions
